requirements.txt
Pillow
requests
exifread